generateBaseMatrix Docstring
Previously I intended to write a docstring for generateBaseMatrix in
each subclass, but the assumptions are the same irrespective of the
longitudinal model, so all the descriptions would all be same. To avoid
duplication, I've moved the docstring to the base class which then gets
inherited by all subclasses (airbag, NHT, arbitraryLongitudinal).

This might be misleading, since not all methods are implemented for all
longitudinal models, but an error will be shown if an unavailable
method is used.

In future, more of this function could be moved to the base class
since all the methods are very similar. It would then be easy to have
a placeholder method that just printed a 'not implemented' message.