Airbag perturbed-full and perturbed-simple methods
The perturbed-full method requires that
w' = w'(l) = p*w_{0} + w_{b} + l*w_{s}

where p is an integer approaching +-infty. If the l*w_{s} term is
neglected (as in perturbed-simple), then w' is a 1D array of values of
frequencies for different values of p. If l*w_{s} is not neglected, as
in the case of perturbed-full, then w' must be a 2D array of values
for different values of l and p, something like this

w' =
[
[ w'(l= 0, p=-1000), w'(l= 0, p=-999), ..., w'(l= 0, p=1000) ],
[ w'(l=+1, p=-1000), w'(l=+1, p=-999), ..., w'(l=+1, p=1000) ],
[ w'(l=+2, p=-1000), w'(l=+2, p=-999), ..., w'(l=+2, p=1000) ],
[ w'(l=-2, p=-1000), w'(l=-2, p=-999), ..., w'(l=-2, p=1000) ],
[ w'(l=-1, p=-1000), w'(l=-1, p=-999), ..., w'(l=-1, p=1000) ]
]

so that when indexed,

w'[ 0] = [ w'(l= 0, p=-1000), w'(l= 0, p=-999), ..., w'(l= 0, p=1000) ]
w'[-1] = [ w'(l=-1, p=-1000), w'(l=-1, p=-999), ..., w'(l=-1, p=1000) ]

etc. Perturbed-simple just has w'[0] as a single 1D array.

For the Bessel functions we need to calculate J_{l}(w'), so for
perturbed-simple we just need a 2D array (or a dictionary as I used
before) so that if we used an array it'd be

[
[J_{ 0}(p=-1000), [J_{ 0}(p=-999), ..., [J_{ 0}(p=+1000)],
[J_{+1}(p=-1000), [J_{+1}(p=-999), ..., [J_{+1}(p=+1000)],
[J_{+2}(p=-1000), [J_{+2}(p=-999), ..., [J_{+2}(p=+1000)],
[J_{-1}(p=-1000), [J_{-1}(p=-999), ..., [J_{-1}(p=+1000)],
[J_{-2}(p=-1000), [J_{-2}(p=-999), ..., [J_{-2}(p=+1000)]
]

For perturbed-full J_{l'}(w'(l)*z/c) has to be calculated
for all l' and l, meaning that a 3D array is required, for example

jdict =
[
 [
  [J_{ 0}(l=0, p=-100), [J_{ 0}(l=0, p=-99), ..., [J_{ 0}(l=0, p=+100)],
  [J_{+1}(l=0, p=-100), [J_{+1}(l=0, p=-99), ..., [J_{+1}(l=0, p=+100)],
  [J_{+2}(l=0, p=-100), [J_{+2}(l=0, p=-99), ..., [J_{+2}(l=0, p=+100)],
  [J_{-1}(l=0, p=-100), [J_{-1}(l=0, p=-99), ..., [J_{-1}(l=0, p=+100)],
  [J_{-2}(l=0, p=-100), [J_{-2}(l=0, p=-99), ..., [J_{-2}(l=0, p=+100)]
 ],
 [
  [J_{ 0}(l=1, p=-100), [J_{ 0}(l=1, p=-99), ..., [J_{ 0}(l=1, p=+100)],
  [J_{+1}(l=1, p=-100), [J_{+1}(l=1, p=-99), ..., [J_{+1}(l=1, p=+100)],
  [J_{+2}(l=1, p=-100), [J_{+2}(l=1, p=-99), ..., [J_{+2}(l=1, p=+100)],
  [J_{-1}(l=1, p=-100), [J_{-1}(l=1, p=-99), ..., [J_{-1}(l=1, p=+100)],
  [J_{-2}(l=1, p=-100), [J_{-2}(l=1, p=-99), ..., [J_{-2}(l=1, p=+100)]
 ],
 [
  [J_{ 0}(l=2, p=-100), [J_{ 0}(l=2, p=-99), ..., [J_{ 0}(l=2, p=+100)],
  [J_{+1}(l=2, p=-100), [J_{+1}(l=2, p=-99), ..., [J_{+1}(l=2, p=+100)],
  [J_{+2}(l=2, p=-100), [J_{+2}(l=2, p=-99), ..., [J_{+2}(l=2, p=+100)],
  [J_{-1}(l=2, p=-100), [J_{-1}(l=2, p=-99), ..., [J_{-1}(l=2, p=+100)],
  [J_{-2}(l=2, p=-100), [J_{-2}(l=2, p=-99), ..., [J_{-2}(l=2, p=+100)]
 ],
 [
  [J_{ 0}(l=-2, p=-100), [J_{ 0}(l=-2, p=-99), ..., [J_{ 0}(l=-2, p=+100)],
  [J_{+1}(l=-2, p=-100), [J_{+1}(l=-2, p=-99), ..., [J_{+1}(l=-2, p=+100)],
  [J_{+2}(l=-2, p=-100), [J_{+2}(l=-2, p=-99), ..., [J_{+2}(l=-2, p=+100)],
  [J_{-1}(l=-2, p=-100), [J_{-1}(l=-2, p=-99), ..., [J_{-1}(l=-2, p=+100)],
  [J_{-2}(l=-2, p=-100), [J_{-2}(l=-2, p=-99), ..., [J_{-2}(l=-2, p=+100)]
 ],
 [
  [J_{ 0}(l=-1, p=-100), [J_{ 0}(l=-1, p=-99), ..., [J_{ 0}(l=-1, p=+100)],
  [J_{+1}(l=-1, p=-100), [J_{+1}(l=-1, p=-99), ..., [J_{+1}(l=-1, p=+100)],
  [J_{+2}(l=-1, p=-100), [J_{+2}(l=-1, p=-99), ..., [J_{+2}(l=-1, p=+100)],
  [J_{-1}(l=-1, p=-100), [J_{-1}(l=-1, p=-99), ..., [J_{-1}(l=-1, p=+100)],
  [J_{-2}(l=-1, p=-100), [J_{-2}(l=-1, p=-99), ..., [J_{-2}(l=-1, p=+100)]
 ]
]

Perturbed-simple is then just jdict[0]. To make this 3D matrix
previously, I changed generateBesselJDict to be a 3-level for loop; one
level is a [for i in x] type loop. This worked, but it made the
function harder to understand and broke the perturbed-simple methods.

In this commit I have changed generateBesselJDict back to how it was
before, but I've left impedanceSampleFrequencies() as it was, so that
there are redundant impedance evaluations if perturbed-simple is used.
In generateSimpleBaseMatrix I have then just indexed the frequency with
l=0, which then makes all the perturbed-simple methods work exactly as
they did before the changes. In generateFullBaseMatrix I've just
run generateBesselJDict in a loop over w, which is then a loop over l
(in the order l=0,1,2,-2,-1) so jdict has an index corresponding to l.

I ran Benchmarks to verify that it all looks right. In most cases I've
kept perturbed-simple for the airbags for speed, but I've also made a
perturbed-full example on one of the A. Burov comparisons. Since this
is a Q=1 resonator at 1.3 GHz, I didn't expect it would make much
difference and the results seem to align very well.