Speed up Laguerre Polynomial Calculation
With the previous method I was using a power operator (**), however
this was always the power of an integer which rose with the iteration,
so a multiplication is faster.

I've also noted an alternative implementation of Lna to the comments
for reference. It seems to work better for large n and alpha, but
struggles when the result is small. It is slow, but can be sped up
massively with lru_cache.