Remove unused multi_bunch_mode parameter
When summing over frequencies the multi-bunch mode changes the
frequencies that are sampled, but when integrating over omega' all
values of omega' are used and the coupled-bunch mode isn't significant.

From Chao's equation 6.236
omega' = (M*p + mu)w_0 + Omega
where mu is the coupled-bunch mode.
When converting the summation to an integral it is assumed that the
summation index p is continuous and the derivative d(omega')/dp used,
which is equal to M*w_0. This adds an extra factor of 1/(M*w_0) to the
coefficient of the interaction matrix. Notably, the multi-bunch mode
is not included.

This factor of 1/M for the broadband case cancels with the extra factor
of M is added to the summation for multiple bunches as per Chao's
Equation 6.237; which is added in generateInteractionMatrix.

The conclusion is that there can be no multi-bunch coupling in the
broadband approximation; all wakes are assumed to die out before the
next bunch arrives. Each bunch only interacts with itself.

Benchmarks all look OK.