Add Code
Pure python files were used to separate the fundamental calculations
from the notebooks, standardise the methods for comparisons between
models, and allow optional JIT compilation with Numba.

Whilst having all functionality in one notebook is good in some ways
(e.g. it's easy to share), it also means that updates are not easy to
pull in. With the new style, the core functions can be updated without
having to modify the notebooks which contain results.

Comparing results from different models has been made easier, by
introducing the class structure and standardising the method names.
This could also help with writing unit tests in the future.

Numba has been found to speed up computation, however it is a large
dependency and can be more challenging to install on some systems. To
ensure the longevity of the code, the Numba dependence should be
optional.